Switch back to load the sg module via the kmod builtin
The problem was not that the kmod builtin is faster then modprobe but
rather the incorrect usage of the "=" assignment operator. We need to
use "+=" here, so the sg module is loaded in addition to other scsi
modules, which are loaded via the modalias rule.
Thanks to Tommaso Colombo for the analysis.