Fix boot failures if CONFIG_UEVENT_HELPER is disabled
Drop the obsolete check for /sys/kernel/uevent_helper from postinst and
the SysV init script and do not unconditionally overwrite it in the
initramfs hook. Since a long time now udev has been using the netlink
interface to communicate with the kernel and with Linux 3.16 it is
possible to disable CONFIG_UEVENT_HELPER completely.

Closes: #752742