Correct udevlibexecdir on multiarch installs.
By default for eudev, the rootlibexecdir is /lib/udev, which is
where arch-specific libraries and config files are installed.
However, this path can be (and is) configured, and all files are
being installed to an arch-specific directory.

Follow the practice of systemd and install the udev config files
to /lib/udev specifically and the arch-specific files to the
configured directory.